Kafka 활용 게임 결과 업데이트 포인트 업데이트 기능 추가

+++ betting-service/src/main/java/com/oneonone/bettingservice/presentation/dto/BettingRequestDto.java
@@ -10,7 +10,7 @@ public record  BettingRequestDto(
         UUID bettingId,
         Long userId,
         UUID gameId,
-        Integer betAmount,
+        BigDecimal betAmount,
         BigDecimal odds,
         BetType betType,
         BetResult betResult
